Move compilers into new module
This also gives me a chance to remove that hacky `hiding
(pandocCompiler)` bit, which is a good change.
